+++ src/screens.rs
@@ -46,6 +46,9 @@ impl Screen {
         screen.read_size(con)?;
         Ok(screen)
     }
+    ///
+    /// Note: with the resize event we could theoretically avoid to call
+    /// this so often
     pub fn read_size(&mut self, con: &AppContext) -> Result<(), ProgramError> {
         let (w, h) = termimad::terminal_size();
         self.width = w;
@@ -53,7 +56,6 @@ impl Screen {
         if let Some(h) = con.launch_args.height {
             self.height = h;
         }
-        debug!("screen size: {} x {}", self.width, self.height);
         self.input_field.change_area(0, h-1, w - FLAGS_AREA_WIDTH);
         Ok(())
     }